{"types":"C","runs":30}

+++ Review.java
@@ -3,7 +3,7 @@ import acm.program.ConsoleProgram;
 public class Review extends ConsoleProgram{
 	public void run(){
 		int a = 5;
-		double b = a / 2;
+		double  b = (double)a / 2;
 		println(b);
 	}
 }
